# Migrate to higher performance bcrypt implement