… packages/strapi-admin/package.json
… yarn.lock