@alexandrebodin 2020-09-18
Hello, Thank you for this PR but we are going to stick with the current library we are using. Performance only is not enough to make us switch to a new library that concerns security :) 

Closing this for now