@Brooooooklyn 2021-01-10
> Performance only is not enough to make us switch to a new library that concerns security :)

`@node-rs/bcrypt` is based on [bcrypt](https://crates.io/crates/bcrypt), which is wildly used in Rust ecosystem (`205k` downloads) and 0 line `unsafe` codes. I believe it is safer and faster than `c` implementation.

BTW, `@node-rs/bcrypt@1.0.0` now support `apple silicon, android aarch64, linux aarch64, linux armv7, linux musl` prebuild binaries, and I believe it will make it easier to install and use.