@Brooooooklyn 2021-02-01
Hey all, `parse-server` has introduced `@node-rs/bcrypt` more than half a year ago, and zero securely issue reported, would you reconsider to introduce `@node-rs/bcrypt` in `strapi`?

Pull request which introduced `@node-rs/bcrypt` into `parser-server`:  https://github.com/parse-community/parse-server/pull/6794

@alexandrebodin